Move code into packages
