+++ src/main/java/luna/command/Command.java
@@ -1,9 +1,9 @@
-package command;
+package luna.command;
 
-import exception.LunaException;
-import storage.Storage;
-import task.TaskList;
-import ui.Ui;
+import luna.exception.LunaException;
+import luna.storage.Storage;
+import luna.task.TaskList;
+import luna.ui.Ui;
 
 import java.io.IOException;
