+++ src/main/java/luna/command/EventCommand.java
@@ -1,9 +1,9 @@
-package command;
+package luna.command;
 
-import storage.Storage;
-import task.Event;
-import task.TaskList;
-import ui.Ui;
+import luna.storage.Storage;
+import luna.task.Event;
+import luna.task.TaskList;
+import luna.ui.Ui;
 
 import java.time.LocalDate;
