+++ src/main/java/luna/command/MarkCommand.java
@@ -1,19 +1,19 @@
-package command;
+package luna.command;
 
-import storage.Storage;
-import task.TaskList;
-import ui.Ui;
+import luna.storage.Storage;
+import luna.task.TaskList;
+import luna.ui.Ui;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private final int taskNumber;
 
-    public UnmarkCommand(int taskNumber) {
+    public MarkCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage<TaskList> storage) {
-        ui.show(taskList.unmarkAsDone(taskNumber));
+        ui.show(taskList.markAsDone(taskNumber));
         saveTaskList(taskList, storage);
     }
