+++ src/main/java/luna/command/UnmarkCommand.java
@@ -1,19 +1,19 @@
-package command;
+package luna.command;
 
-import storage.Storage;
-import task.TaskList;
-import ui.Ui;
+import luna.storage.Storage;
+import luna.task.TaskList;
+import luna.ui.Ui;
 
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
     private final int taskNumber;
 
-    public MarkCommand(int taskNumber) {
+    public UnmarkCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage<TaskList> storage) {
-        ui.show(taskList.markAsDone(taskNumber));
+        ui.show(taskList.unmarkAsDone(taskNumber));
         saveTaskList(taskList, storage);
     }
