+++ src/main/java/luna/exception/LunaException.java
@@ -1,4 +1,4 @@
-package exception;
+package luna.exception;
 
 public class LunaException extends RuntimeException {
     public LunaException(String errorMessage) {
