+++ src/main/java/luna/task/Deadline.java
@@ -1,4 +1,4 @@
-package task;
+package luna.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
