+++ src/main/java/luna/task/Task.java
@@ -1,6 +1,6 @@
-package task;
+package luna.task;
 
-import exception.LunaException;
+import luna.exception.LunaException;
 
 import java.io.Serializable;
